+++ package/thunderstore/BiomeGate/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+* debuff will not apply while player is in cutscene to prevent debuff messages showing while in intro
+* option to hide icon from status effects list
+* minor refinements
+
 # 1.0.1
 * option to disable particular biome gating if global key is set